Introduce the FileRunner framework in libqtwine.
The basic idea behind this framework is the need for a way
to handle various different file types that wine can handle.
With this framework we have a FileRunner class that receives a
filename and some options and internally launches a plugin that
will handle the specified file type in an appropriate way.
It is designed in this way so that it is easy to subclass either
FileRunner or one of its plugins and add extra functionality like
dialogs to manage the options, etc...

This commit adds the base classes (FileRunner and FileRunnerPlugin)
and two plugins: one to execute .exe files and one to merge .reg files
in the registry.